src/ragcopy/requirements.txt
python-irodsclient>=1.1.8
